+++ src/main/resources/static/js/site/timeline.js
@@ -86,12 +86,11 @@
                 }
             }
 
-            if (Util.isSafari() === true || Util.isFirefox() === true) {
-                elements.each(function (index) {
-                    let id = $(this).attr("id");
+            elements.each(function (index) {
+                let id = $(this).attr("id");
 
-                    if (id.indexOf("tail_") === -1 && index < 2 && timelineSettings.prevAnchor !== id) {
-
+                if (id.indexOf("tail_") === -1 && index < 2 && timelineSettings.prevAnchor !== id) {
+                    if (Util.isSafari() === true || Util.isFirefox() === true) {
                         timelineSettings.renderThumbnails(id, mediaTypeFilter).then(function (msg) {
                             if (msg === timelineSettings.successBelowMsg || msg === timelineSettings.successAboveMsg || msg === timelineSettings.successMidMsg) {
                                 timelineSettings.setScrollSpyActive(id);
@@ -99,8 +98,18 @@
                         });
                         timelineSettings.prevAnchor = id;
                     }
-                });
-            } else {
+
+                    // Set the timeline slider while scrolling
+                    timelineDates.forEach(function (timelineDate, i) {
+                        if (id === timelineDate.year + "-" + timelineDate.month + "-" + timelineDate.day) {
+                            $("#dateSlider").slider("option", "value", timelineDates.length-i);
+                            return false;
+                        }
+                    });
+                }
+            });
+
+            if (Util.isChrome() === true) {
                 timelineSettings.renderThumbnailsSimple(elements, mediaTypeFilter, timelineDates).then(function (msg) {
                     if (msg === timelineSettings.success) {
                         // Set TOC active element
@@ -507,8 +516,100 @@
         return timelineSettings.successMidMsg;
     }
 
+    timelineSettings.initializeTimelineSlider = function (dateList, mediaTypeFilter) {
+        $("#dateSlider").hide();
+
+        // Tooltip for handle
+        const handleTooltip = $('<span class="badge bg-light text-dark" id="tooltip" />').css({
+            position: 'absolute',
+            right: 12
+        }).hide();
+
+        handleTooltip.text(Util.getShortMonths(dateList[0].month-1) + ' ' + dateList[0].year);
+
+        $("#dateSlider").slider({
+            orientation: "vertical",
+            value: dateList.length-1,
+            min: 0,
+            max: dateList.length - 1,
+            step : 0.0001,
+            range: false,
+            slide: function(event, ui) {
+                const currentDateObj = dateList[Math.round((dateList.length - 1)-ui.value)];
+                handleTooltip.text(Util.getShortMonths(currentDateObj.month-1) + ' ' + currentDateObj.day + ', ' + currentDateObj.year);
+            },
+            stop: function(event, ui) {
+                const currentDateObj = dateList[Math.round((dateList.length - 1)-ui.value)];
+
+                if (timelineSettings.enableScrollSpy === true) {
+                    timelineSettings.jumpFromTimelineToc(event, currentDateObj.year + '-' + currentDateObj.month + '-' + currentDateObj.day, mediaTypeFilter);
+                }
+            }
+        }).find(".ui-slider-handle").append(handleTooltip).hover(function() {
+            handleTooltip.show();
+        }, function() {
+            handleTooltip.hide();
+        });
+
+        // Render ticks
+        for (let i = 0; i <= dateList.length; i++) {
+            const timelineDateObj = dateList[i];
+            if (timelineDateObj) {
+                const dateObj = new Date(timelineDateObj.month + "/" + timelineDateObj.day + "/" + timelineDateObj.year)
+                if (i === 0 || i > 0 && dateList[i - 1].year !== timelineDateObj.year) {
+                    // Label for year
+                    const el = $('<span class="badge rounded-pill bg-light text-dark">' + dateObj.getFullYear() + '</span>').css({
+                        'width': '35px',
+                        'right': '15px',
+                        'position':'absolute',
+                        'top':(i / dateList.length * 100) + '%'
+                    });
+                    $("#dateSlider").append(el);
+                } else if (i === 0 || (i > 0 && (dateList[i - 1].year !== timelineDateObj.year || dateList[i - 1].month !== timelineDateObj.month))) {
+                    // Tick for month/year
+                    const tickEl = $('<span data-tick-id="'+timelineDateObj.year + '-' + timelineDateObj.month +'">' + '-' + '</span>').css({
+                        'width': '10px',
+                        'right': '15px',
+                        'position':'absolute',
+                        'top':(i / dateList.length * 100) + '%'
+                    });
+                    $("#dateSlider").append(tickEl);
+                }
+
+                // Tooltip for month/year on slider
+                const sliderTooltip = $('<span class="badge bg-light text-dark" id="tooltip" />').css({
+                    position: 'absolute',
+                    right: 12
+                }).hide();
+                sliderTooltip.text(Util.getShortMonths(timelineDateObj.month-1) + ' ' + timelineDateObj.year);
+
+                const sliderEl = $('<span data-slider-id="'+timelineDateObj.year + '-' + timelineDateObj.month +'">&nbsp;</span>').css({
+                    'width': '70px',
+                    'right': '0px',
+                    // 'background-color': 'grey',
+                    'position':'absolute',
+                    'top':(i / dateList.length * 100) + '%'
+                });
+                $(sliderEl).append(sliderTooltip);
+                $("#dateSlider").append(sliderEl);
+
+                sliderEl.hover(function() {
+                    sliderTooltip.show();
+                }, function() {
+                    sliderTooltip.hide();
+                });
+            }
+        }
+
+        $("#dateSliderWrapper").hover(function() {
+            $("#dateSlider").show();
+        }, function() {
+            $("#dateSlider").hide();
+        });
+    }
+
     timelineSettings.jumpFromTimelineToc = function (e,anchor,mediaTypeFilter) {
-        e.preventDefault();
+        //e.preventDefault();
 
         timelineSettings.currentScrollDirection = timelineSettings.ScrollDirection.down;
         timelineSettings.enableScrollSpy = false;