+++ CHANGES.md
@@ -38,5 +38,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Simplified timeline view logic
 - Optimized modals on views
+- Rehauled date scrolling
 
 ### Fixed